wip LIDAR safety
